requirements.txt
python-binance
pandas
Flask
python-telegram-bot[job-queue]

